Address https://errorprone.info/bugpattern/UnnecessaryParentheses warnings in tests.
A couple notes on the `primitives` changes:
- I realized that we might as well use `Double.hashCode(double)` and `Float.hashCode(float)` there instead of boxing and calling the instance `hashCode()` method. (I'm going to change the _prod_ code in another CL.)
- We probably don't even need to make sure that the expected values in our assertions have the exact right runtime type, given that Truth considers, e.g., the `Integer` value `1` to be equal to the `Long` value `1L`. I still kept the types matching when it was easy to do (e.g., `1.0`, `1.0f`), but I didn't bother when it requires a cast (`(short) 1`).

PiperOrigin-RevId: 745662180

+++ android/guava-tests/test/com/google/common/primitives/LongsTest.java
@@ -578,14 +578,14 @@ public class LongsTest extends TestCase {
 
   public void testStringConverter_convert() {
     Converter<String, Long> converter = Longs.stringConverter();
-    assertThat(converter.convert("1")).isEqualTo((Long) 1L);
-    assertThat(converter.convert("0")).isEqualTo((Long) 0L);
-    assertThat(converter.convert("-1")).isEqualTo((Long) (-1L));
-    assertThat(converter.convert("0xff")).isEqualTo((Long) 255L);
-    assertThat(converter.convert("0xFF")).isEqualTo((Long) 255L);
-    assertThat(converter.convert("-0xFF")).isEqualTo((Long) (-255L));
-    assertThat(converter.convert("#0000FF")).isEqualTo((Long) 255L);
-    assertThat(converter.convert("0666")).isEqualTo((Long) 438L);
+    assertThat(converter.convert("1")).isEqualTo(1L);
+    assertThat(converter.convert("0")).isEqualTo(0L);
+    assertThat(converter.convert("-1")).isEqualTo(-1L);
+    assertThat(converter.convert("0xff")).isEqualTo(255L);
+    assertThat(converter.convert("0xFF")).isEqualTo(255L);
+    assertThat(converter.convert("-0xFF")).isEqualTo(-255L);
+    assertThat(converter.convert("#0000FF")).isEqualTo(255L);
+    assertThat(converter.convert("0666")).isEqualTo(438L);
   }
 
   public void testStringConverter_convertError() {
